wandb/run-20250710_121454-ddbbsncb/files/requirements.txt
iniconfig==2.0.0
albumentations==2.0.3
triton==3.1.0
aiohttp==3.11.11
joblib==1.4.2
omegaconf==2.3.0
rich==13.9.4
fsspec==2024.2.0
annotated-types==0.7.0
wrapt==1.17.2
flatbuffers==25.1.24
google-pasta==0.2.0
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
torchtyping==0.1.5
wheel==0.44.0
pluggy==1.5.0
soupsieve==2.6
optree==0.14.0
debugpy==1.8.12
nvidia-cublas-cu12==12.1.3.1
mpmath==1.3.0
torchvision==0.20.0+cu121
torch_cluster==1.6.3+pt25cu121
RapidFuzz==3.13.0
GitPython==3.1.44
tqdm==4.67.1
jaxtyping==0.2.36
parmap==1.7.0
umap==0.1.1
bs4==0.0.2
opencv-python-headless==4.11.0.86
ipykernel==6.29.5
aiohappyeyeballs==2.4.4
petname==2.6
attrs==24.3.0
black==25.1.0
filelock==3.13.1
imageio==2.37.0
ipython==8.31.0
typeguard==2.13.3
pyg-lib==0.4.0+pt25cu121
keras==3.8.0
pyzmq==26.2.0
safetensors==0.5.2
torch_sparse==0.6.18+pt25cu121
networkx==3.2.1
charset-normalizer==3.4.1
nvidia-cuda-runtime-cu12==12.1.105
ptyprocess==0.7.0
torch_scatter==2.1.2+pt25cu121
comm==0.2.2
pandas==2.2.3
pydantic_core==2.27.2
nvidia-cuda-cupti-cu12==12.1.105
jax==0.5.0
contourpy==1.3.1
python-dateutil==2.9.0.post0
regex==2024.11.6
scikit-learn==1.6.1
cycler==0.12.1
stringzilla==3.11.3
Markdown==3.7
executing==2.1.0
kiwisolver==1.4.8
pickleshare==0.7.5
fasttext==0.9.3
torchinfo==1.8.0
fonttools==4.55.3
jiwer==3.1.0
Pygments==2.19.1
dill==0.3.9
albucore==0.0.23
ultralytics-thop==2.0.14
editdistance==0.8.1
importlib_metadata==8.5.0
jedi==0.19.2
setproctitle==1.3.4
astunparse==1.6.3
aiosignal==1.3.2
pydantic==2.10.5
psutil==6.1.1
timm==1.0.13
jaxlib==0.5.0
zipp==3.21.0
markdown-it-py==3.0.0
pox==0.3.5
termcolor==2.5.0
tomli==2.2.1
nvidia-cusparse-cu12==12.1.0.106
yarl==1.18.3
Unidecode==1.3.8
pip==24.2
ppft==1.7.6.9
asttokens==3.0.0
torch_pca==1.1.0
traitlets==5.14.3
nest_asyncio==1.6.0
absl-py==2.1.0
scipy==1.15.1
libclang==18.1.1
sympy==1.13.1
opencv-contrib-python==4.11.0.86
wcwidth==0.2.13
py-cpuinfo==9.0.0
protobuf==4.25.6
simsimd==6.2.1
multiprocess==0.70.17
torch-geometric==2.6.1
six==1.17.0
einops==0.8.0
matplotlib==3.10.0
pathspec==0.12.1
nvidia-nvjitlink-cu12==12.1.105
pexpect==4.9.0
lightning-utilities==0.12.0
opt_einsum==3.4.0
pure_eval==0.2.3
torchaudio==2.5.0+cu121
sentry-sdk==2.20.0
hydra-core==1.3.2
lazy_loader==0.4
jupyter_client==8.6.3
packaging==24.2
hf-xet==1.1.0
gitdb==4.0.12
propcache==0.2.1
nvidia-cuda-nvrtc-cu12==12.1.105
pillow==11.2.1
transformers==4.51.1
torch==2.5.0+cu121
threadpoolctl==3.5.0
gast==0.6.0
mdurl==0.1.2
tensorboard-data-server==0.7.2
matplotlib-inline==0.1.7
Jinja2==3.1.3
seaborn==0.13.2
PyYAML==6.0.2
nvidia-cusolver-cu12==11.4.5.107
MarkupSafe==2.1.5
click==8.1.8
nvidia-nccl-cu12==2.21.5
tokenizers==0.21.0
urllib3==2.3.0
jupyter_core==5.7.2
antlr4-python3-runtime==4.9.3
prompt_toolkit==3.0.48
nvidia-ml-py==12.560.30
stack_data==0.6.3
smmap==5.0.2
pyparsing==3.2.1
pytest==8.3.4
pathos==0.3.3
cachetools==5.5.0
nvidia-cudnn-cu12==9.1.0.70
wandb==0.19.10
namex==0.0.8
docker-pycreds==0.4.0
ctc-decoder==1.0.1
async-timeout==5.0.1
platformdirs==4.3.6
exceptiongroup==1.2.2
tifffile==2025.3.30
idna==3.10
word-beam-search==1.0.1
torch_spline_conv==1.2.2+pt25cu121
h5py==3.12.1
multidict==6.1.0
tzdata==2024.2
torchdiffeq==0.2.5
nvidia-cufft-cu12==11.0.2.54
ml-dtypes==0.3.2
tornado==6.4.2
torchmetrics==1.6.1
certifi==2024.12.14
numpy==1.26.3
huggingface-hub==0.30.2
parso==0.8.4
tensorflow-io-gcs-filesystem==0.37.1
scikit-image==0.25.2
beautifulsoup4==4.12.3
nvidia-curand-cu12==10.3.2.106
ultralytics==8.3.64
Werkzeug==3.1.3
pybind11==2.13.6
tensorboard==2.16.2
decorator==5.1.1
nvitop==1.4.1
typing_extensions==4.12.2
requests==2.32.3
pytorch-metric-learning==2.8.1
setuptools==75.1.0
frozenlist==1.5.0
pytz==2024.2
mypy-extensions==1.0.0
grpcio==1.70.0
jaraco.collections==5.1.0
backports.tarfile==1.2.0
typeguard==4.3.0
jaraco.functools==4.0.1
packaging==24.1
importlib_metadata==8.0.0
jaraco.text==3.12.1
more-itertools==10.3.0
autocommand==2.2.2
inflect==7.3.1
zipp==3.19.2
wheel==0.43.0
jaraco.context==5.3.0
importlib_resources==6.4.0
typing_extensions==4.12.2
tomli==2.0.1
platformdirs==4.2.2
